+++ commands/ping.js
@@ -1,0 +1,10 @@
+var matcher = /!ping/;
+
+function run(command, request) {
+  return {
+    'text': 'pong'
+  };
+}
+
+exports.run = run;
+exports.matcher = matcher;
